new changes

+++ src/main/java/local/backend/daos/BooksDAO.java
@@ -4,4 +4,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import local.backend.entity.Books;
 
-public interface BooksDAO extends JpaRepository<Books, Long> {  }
+public interface BooksDAO extends JpaRepository<Books, Integer> {  }
